Moved commands to new directory
